Permisos para ingreso a rutas
se implementaron los permisos para  cada una de las rutas

+++ app/Http/Controllers/api/NoteController.php
@@ -4,32 +4,112 @@ namespace App\Http\Controllers\api;
 
 use App\Http\Controllers\Controller;
 use App\Models\Note;
-use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
-
-
+use Illuminate\Support\Facades\Redis;
 
 class NoteController extends Controller
 {
+    // Tipos de usuario Bases
+    const rolMaster = 1;
+    const rolAdmin = 2;
+    const rolUsuario = 3;
 
 
-    // Controllador Index de Notes -------------------------------------------------------------------------------------
+
+
+    // Controllador Index de Notes (trae solo las notas del usuario)-------------------------------------------------------------------------------------
     public function index()
     {
+        try {
 
-        $notes = Note::all();
-        return response()->json($notes);
+            $user = Auth::user();
+            $notes = Note::where('user_id', $user->id)
+                ->join('states', 'notes.state_id', '=', 'states.id')
+                ->join('visibilities', 'notes.visibility_id', '=', 'visibilities.id')
+                ->select('notes.id', 'notes.title', 'notes.text', 'states.state_name', 'visibilities.visibility_name')
+                ->get();
+
+
+            // manejo de ausencia notas
+            if ($notes->isEmpty()) {
+                return response()->json(
+                    [
+                        'status' => true,
+                        'message' => 'No tienes notas'
+                    ],
+                    200
+                );
+            } else {
+                return response()->json(
+                    [
+                        'status' => true,
+                        'notes' => $notes
+                    ]
+                );
+            }
+        } catch (\Throwable $ex) {
+
+            //registro error Log
+            Log::error('Error en NoteController@index: ' . $ex->getMessage());
+
+            return response()->json(
+                [
+                    'status' => false,
+                    'message' => 'Se produjo un error en el servidor'
+                ],
+                500
+            );
+        }
     }
 
 
 
-    // Controllador Create de Notes (ruta para crear nuevas notas) ----------------------------------------------------
-    public function create()
+
+    // Controllador publicNote de Notes (trae todas las notas que sean publicas)-------------------------------------------------------------------------
+    public function publicNote()
     {
-        return view('crearNota');
-    }
+        try {
+            $notes = DB::table('notes')
+                ->join('users', 'notes.user_id', '=', 'users.id')
+                ->join('states', 'notes.state_id', '=', 'states.id')
+                ->where('notes.visibility_id', 1)
+                ->select('notes.title', 'notes.text', 'users.name', 'states.state_name')
+                ->get();
 
+            if ($notes->isEmpty()) {
+                return response()->json(
+                    [
+                        'status' => true,
+                        'message' => 'No hay notas'
+                    ],
+                    200
+                );
+            } else {
+                return response()->json(
+                    [
+                        'status' => true,
+                        'notes' => $notes
+                    ]
+                );
+            }
+        } catch (\Throwable $ex) {
+
+            //registro error Log
+            Log::error('Error en NoteController@publicNote: ' . $ex->getMessage());
+
+            return response()->json(
+                [
+                    'status' => false,
+                    'message' => 'Se produjo un error en el servidor'
+
+                ],
+                500
+            );
+        }
+    }
 
 
     // Controllador Store de Notes (Crea nota) ------------------------------------------------------------------------
@@ -40,32 +120,40 @@ class NoteController extends Controller
             $validateNote = $request->validate([
                 'title' => 'required|string|max:50',
                 'text' => 'required|string|max:255',
-                'user_id' => 'required|integer|exists:users,id',
                 'state_id' => 'required|integer|exists:states,id',
                 'visibility_id' => 'required|integer|exists:visibilities,id'
-
             ]);
+
+            // Incertar Id  usuario autenticado como  User_id
+            $validateNote['user_id'] = Auth::id();
 
             $note = Note::create($validateNote);
 
             return response()->json(
                 [
+                    'status' => true,
                     'message' => 'Nota creada exitosamente',
-                    'note' => $note
+                    'note' => [
+                        'title' => $note['title'],
+                        'text' => $note['text'],
+                        'state_id' => $note['state_id'],
+                        'visibility_id' => $note['visibility_id']
+                    ]
                 ],
                 200
             );
 
             // Catch errores
-        } catch (Exception $error) {
-            // Registra  error
-            Log::error('Error al crear la nota: ' . $error->getMessage());
+        } catch (\Throwable $ex) {
 
-            // Respuesta error
+            //registro error Log
+            Log::error('Error en NoteController@store: ' . $ex->getMessage());
+
             return response()->json(
                 [
-                    'mensaje' => 'Error al crear la nota',
-                    'error' => $error->getMessage()
+                    'status' => false,
+                    'message' => 'Se produjo un error en el servidor'
+
                 ],
                 500
             );
@@ -74,31 +162,112 @@ class NoteController extends Controller
 
 
 
+
+
     // Controllador Show de Notes (busca)-------------------------------------------------------------------------------
     public function show($id)
     {
-        
-        $note = Note::find($id);
-        if ($note) {
-            return response()->json($note);
-        } else {
+        try {
+
+            $note = Note::find($id);
+            if ($note) {
+
+                $user = Auth::user();
+                if ($user->id == $note->user_id || $user->rol_id == self::rolAdmin || $user->rol_id == self::rolMaster) {
+                    return response()->json([
+                        'status' => true,
+                        'note' => $note
+                    ]);
+                } else {
+                    return response()->json(
+                        [
+                            'status' => false,
+                            'message' => 'No autorizado'
+                        ],
+                        403
+                    );
+                }
+            } else {
+                return response()->json(
+                    [
+                        'status' => false,
+                        'message' => 'Nota no encontrada'
+                    ],
+                    404
+                );
+            }
+
+            // Catch errores
+        } catch (\Throwable $ex) {
+
+            //registro error Log
+            Log::error('Error en NoteController@show: ' . $ex->getMessage());
+
             return response()->json(
                 [
-                    'message' => 'Nota no encontrada'
+                    'status' => false,
+                    'message' => 'Se produjo un error en el servidor'
                 ],
-                404
+                500
             );
         }
     }
 
 
 
-    // Controllador edit de Notes (ruta para editar notas) -----------------------------------------------------------
-    public function edit()
+    // Controllador Allnote (muestra toda informacion notas a Admons) ----------------------------------------------------------------------------
+    public function allNote()
     {
-        return view('editarNota');
-    }
+        try {
 
+            $user = Auth::User();
+            if ($user->rol_id == self::rolAdmin || $user->rol_id == self::rolMaster) {
+
+                $notes = DB::table('notes')
+                    ->join('users', 'notes.user_id', '=', 'users.id')
+                    ->join('states', 'notes.state_id', '=', 'states.id')
+                    ->select('notes.id', 'notes.user_id', 'users.name', 'states.state_name', 'notes.title', 'notes.text')
+                    ->get();
+
+                if ($notes->isEmpty()) {
+                    return response()->json(
+                        [
+                            'status' => true,
+                            'message' => 'No hay notas'
+                        ],
+                        200
+                    );
+                } else {
+                    return response()->json(
+                        [
+                            'status' => true,
+                            'notes' => $notes
+                        ]
+                    );
+                }
+            } else {
+                return response()->json(
+                    [
+                        'status' => false,
+                        'message' => 'No autorizado'
+                    ],
+                    403
+                );
+            }
+        } catch (\Throwable $ex) {
+
+            //registro error Log
+            Log::error('Error en NoteController@allNote: ' . $ex->getMessage());
+
+            return response()->json(
+                [
+                    'status' => false,
+                    'message' => 'Se produjo un error en el servidor'
+                ],
+                500
+            );
+        }
+    }
 
     // Controllador update de Notes (Actualiza notas) -------------------------------------------------------------------
     public function update(Request $request, $id)
@@ -106,37 +275,55 @@ class NoteController extends Controller
 
         try {
 
+            $user = Auth::user();
             $note = Note::findOrFail($id);
-            //validaciones de la informacion recibida por Json
-            $validateNote = $request->validate([
-                'title' => 'required|string|max:50',
-                'text' => 'required|string|max:255',
-                'user_id' => 'required|integer|exists:users,id',
-                'state_id' => 'required|integer|exists:states,id',
-                'visibility_id' => 'required|integer|exists:visibilities,id'
 
-            ]);
+            if ($user->id == $note->user_id || $user->rol_id == self::rolAdmin || $user->rol_id == self::rolMaster) {
 
-            $note->update($validateNote);
+                //validaciones de la informacion recibida por Json
+                $validateNote = $request->validate([
+                    'title' => 'required|string|max:50',
+                    'text' => 'required|string|max:255',
+                    'state_id' => 'required|integer|exists:states,id',
+                    'visibility_id' => 'required|integer|exists:visibilities,id'
 
-            return response()->json(
-                [
-                    'message' => 'Nota Actualizada exitosamente',
-                    'note' => $note
-                ],
-                200
-            );
+                ]);
 
+                $note->update($validateNote);
+
+                return response()->json(
+                    [
+                        'status' => true,
+                        'message' => 'Nota Actualizada exitosamente',
+                        'note' => [
+                            'title' => $validateNote['title'],
+                            'text' => $validateNote['text'],
+                            'state_id' => $validateNote['state_id'],
+                            'visibility_id' => $validateNote['visibility_id']
+                        ]
+                    ],
+                    200
+                );
+            } else {
+                return response()->json(
+                    [
+                        'status' => false,
+                        'message' => 'No autorizado'
+                    ],
+                    403
+                );
+            }
             // Catch errores
-        } catch (Exception $error) {
-            // Registra  error
-            Log::error('Error al Actualizar la nota: ' . $error->getMessage());
+        } catch (\Throwable $ex) {
 
-            // Respuesta error
+            //registro error Log
+            Log::error('Error en NoteController@update ' . $ex->getMessage());
+
             return response()->json(
                 [
-                    'mensaje' => 'Error al Actualizar la nota',
-                    'error' => $error->getMessage()
+                    'status' => false,
+                    'message' => 'Se produjo un error en el servidor'
+
                 ],
                 500
             );
@@ -151,32 +338,54 @@ class NoteController extends Controller
     {
         try {
             $note = Note::find($id);
+            $user = Auth::user();
 
             if ($note) {
-                $note->delete();
 
-                return response()->json(
-                    [
-                        'message' => 'Nota eliminada exitosamente'
-                    ],
-                    200
-                );
+                //Permiso para eliminar 
+                if ($user->id == $note->user_id || $user->rol_id == self::rolAdmin || $user->rol_id == self::rolMaster) {
+
+                    //Elimina nota
+                    $note->delete();
+
+                    return response()->json(
+                        [
+                            'status' => true,
+                            'message' => 'Nota eliminada exitosamente'
+                        ],
+                        200
+                    );
+                } else {
+
+                    //Respuesta NO autorizado
+                    return response()->json(
+                        [
+                            'status' => false,
+                            'message' => 'No autorizado'
+                        ],
+                        403
+                    );
+                }
             } else {
+
+                //Respuesta  No existe nota
                 return response()->json(
                     [
+                        'status' => false,
                         'message' => 'Nota no encontrada'
                     ],
                     404
                 );
             }
-        } catch (\Exception $error) {
-            // Catch errores
-            Log::error('Error al eliminar la nota: ' . $error->getMessage());
+        } catch (\Throwable $ex) {
+
+            //registro error Log
+            Log::error('Error en NoteController@destroy ' . $ex->getMessage());
 
             return response()->json(
                 [
-                    'mensaje' => 'Error al eliminar la nota',
-                    'error' => $error->getMessage()
+                    'status' => false,
+                    'message' => 'Se produjo un error en el servidor'
                 ],
                 500
             );
