+++ app/Http/Controllers/api/RolController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\api;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Models\Rol;
 use Exception;
 use Illuminate\Http\Request;
@@ -10,21 +11,54 @@ use Illuminate\Support\Facades\Log;
 
 class RolController extends Controller
 {
+    // Tipos de usuario Bases
+    const rolMaster = 1;
+    const rolAdmin = 2;
+    const rolUsuario = 3;
+
+
+
     // Controllador Index de Rols -------------------------------------------------------------------------------------
     public function index()
     {
+        try {
 
-        $rols = Rol::all();
-        return response()->json($rols);
+            // Verificar Rol Admin
+            $rols = Rol::select('id', 'rol_name')->get();
+
+            // No hay roles en BD
+            if ($rols->isEmpty()) {
+                return response()->json(
+                    [
+                        'status' => true,
+                        'message' => 'No hay rols'
+                    ]
+                );
+
+                // Retorna roles de la Bd
+            } else {
+                return response()->json(
+                    [
+                        'status' => true,
+                        'rols' => $rols
+                    ]
+                );
+            }
+        } catch (\Throwable $ex) {
+
+            //registro error Log
+            Log::error('Error en RolController@index ' . $ex->getMessage());
+
+            return response()->json(
+                [
+                    'status' => false,
+                    'message' => 'Se produjo un error en el servidor'
+                ],
+                500
+            );
+        }
     }
 
-
-
-    // Controllador Create de Rols (ruta para crear nuevos roles) ----------------------------------------------------
-    public function create()
-    {
-        return view('crearRol');
-    }
 
 
 
@@ -32,33 +66,46 @@ class RolController extends Controller
     public function store(Request $request)
     {
         try {
-            //validaciones de la informacion recibida por Json
-            $validateRol = $request->validate([
+            $user = Auth::user();
+            if ($user->rol_id == self::rolAdmin || $user->rol_id == self::rolMaster) {
 
-                'rol_name' => 'required|string|max:45|unique:rols,rol_name',
+                //validaciones de la informacion recibida por Json
+                $validateRol = $request->validate(
+                    [
+                        'rol_name' => 'required|string|max:45|unique:rols,rol_name',
+                    ]
+                );
 
-            ]);
+                $rol = Rol::create($validateRol);
 
-            $rol = Rol::create($validateRol);
+                return response()->json(
+                    [
+                        'status' => true,
+                        'message' => 'rol creado exitosamente',
+                        'rol' => $rol
+                    ],
+                    200
+                );
+            } else {
+                return response()->json(
+                    [
+                        'status' => false,
+                        'message' => 'No autorizado'
+                    ],
+                    403
+                );
+            }
 
-            return response()->json(
-                [
-                    'message' => 'rol creado exitosamente',
-                    'rol' => $rol
-                ],
-                200
-            );
 
             // Catch errores
-        } catch (Exception $error) {
+        } catch (Exception $ex) {
             // Registra  error
-            Log::error('Error al crear el rol: ' . $error->getMessage());
+            Log::error('Error en RolController@store ' . $ex->getMessage());
 
-            // Respuesta error
             return response()->json(
                 [
-                    'mensaje' => 'Error al crear el rol',
-                    'error' => $error->getMessage()
+                    'status' => false,
+                    'message' => 'Se produjo un error en el servidor'
                 ],
                 500
             );
@@ -70,27 +117,41 @@ class RolController extends Controller
     // Controllador Show de Rols (busca)-------------------------------------------------------------------------------
     public function show($id)
     {
+        try {
 
-        $rol = Rol::find($id);
-        if ($rol) {
-            return response()->json($rol);
-        } else {
+            $rol = Rol::select('id', 'rol_name', 'updated_at')->find($id);
+            if ($rol) {
+
+                return response()->json([
+                    'status' => true,
+                    'rol' => $rol
+                ]);
+            } else {
+                return response()->json(
+                    [
+                        'status' => false,
+                        'message' => 'Rol no encontrada'
+                    ],
+                    404
+                );
+            }
+
+            // Catch errores
+        } catch (\Throwable $ex) {
+
+            //registro error Log
+            Log::error('Error en RolController@show: ' . $ex->getMessage());
+
             return response()->json(
                 [
-                    'message' => 'Rol no encontrado'
+                    'status' => false,
+                    'message' => 'Se produjo un error en el servidor'
                 ],
-                404
+                500
             );
         }
     }
 
-
-
-    // Controllador edit de Rols (ruta para editar roles) -----------------------------------------------------------
-    public function edit()
-    {
-        return view('editarRol');
-    }
 
 
 
@@ -99,35 +160,55 @@ class RolController extends Controller
     {
 
         try {
-
+            $user = Auth::user();
             $rol = Rol::findOrFail($id);
-            //validaciones de la informacion recibida por Json
-            $validateRol = $request->validate([
+            if ($rol) {
+                if (($user->rol_id == self::rolAdmin and $rol->id != 1 and $rol->id != 2) || $user->rol_id == self::rolMaster) { //Permite modificar solos otros que no sean los 2 iniciales
+                    //validaciones de la informacion recibida por Json
+                    $validateRol = $request->validate([
 
-                'rol_name' => 'required|string|max:45|unique:rols,rol_name,'.$id.',id'
+                        'rol_name' => 'required|string|max:45|unique:rols,rol_name,' . $id . ',id'
 
-            ]);
+                    ]);
 
-            $rol->update($validateRol);
+                    $rol->update($validateRol);
 
-            return response()->json(
-                [
-                    'message' => 'Rol Actualizado exitosamente',
-                    'rol' => $rol
-                ],
-                200
-            );
+                    return response()->json(
+                        [
+                            'message' => 'Rol Actualizado exitosamente',
+                            'rol' => $rol
+                        ],
+                        200
+                    );
+                } else {
+                    return response()->json(
+                        [
+                            'status' => false,
+                            'message' => 'No autorizado'
+                        ],
+                        403
+                    );
+                }
+            } else {
+                return response()->json(
+                    [
+                        'status' => false,
+                        'message' => 'Rol no encontrado'
+                    ],
+                    404
+                );
+            }
 
             // Catch errores
-        } catch (Exception $error) {
-            // Registra  error
-            Log::error('Error al Actualizar el rol: ' . $error->getMessage());
+        } catch (\Throwable $ex) {
 
-            // Respuesta error
+            //registro error Log
+            Log::error('Error en RolController@update: ' . $ex->getMessage());
+
             return response()->json(
                 [
-                    'mensaje' => 'Error al Actualizar el rol',
-                    'error' => $error->getMessage()
+                    'status' => false,
+                    'message' => 'Se produjo un error en el servidor'
                 ],
                 500
             );
@@ -142,32 +223,54 @@ class RolController extends Controller
     {
         try {
             $rol = Rol::find($id);
+            $user = Auth::user();
 
             if ($rol) {
-                $rol->delete();
 
-                return response()->json(
-                    [
-                        'message' => 'Rol eliminado exitosamente'
-                    ],
-                    200
-                );
+                //Permiso para eliminar 
+                if (($user->rol_id == self::rolAdmin and $rol->id != 1 and $rol->id != 2) || $user->rol_id == self::rolMaster) {
+
+                    //Elimina rol
+                    $rol->delete();
+
+                    return response()->json(
+                        [
+                            'status' => true,
+                            'message' => 'Rol eliminado exitosamente'
+                        ],
+                        200
+                    );
+                } else {
+
+                    //Respuesta NO autorizado
+                    return response()->json(
+                        [
+                            'status' => false,
+                            'message' => 'No autorizado'
+                        ],
+                        403
+                    );
+                }
             } else {
+
+                //Respuesta  No existe rol
                 return response()->json(
                     [
-                        'message' => 'rol no encontrado'
+                        'status' => false,
+                        'message' => 'Rol no encontrado'
                     ],
                     404
                 );
             }
-        } catch (\Exception $error) {
-            // Catch errores
-            Log::error('Error al eliminar el rol: ' . $error->getMessage());
+        } catch (\Throwable $ex) {
+
+            //registro error Log
+            Log::error('Error en RolController@destroy ' . $ex->getMessage());
 
             return response()->json(
                 [
-                    'mensaje' => 'Error al eliminar el rol',
-                    'error' => $error->getMessage()
+                    'status' => false,
+                    'message' => 'Se produjo un error en el servidor'
                 ],
                 500
             );
